refactor: move contacto page content to component to fix encoding
